pyproject.toml
[tool.poetry]
name = "aien-api"
version = "0.1.0"
description = ""
authors = ["Koeritz, Lisa <lisa.koeritz@uni-tuebingen.de>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.114.1"
modal = "^0.64.101"
pydantic = "^2.9.1"
qdrant-client = "^1.11.1"
python-decouple = "^3.8"
langchain-qdrant = "^0.1.3"
uvicorn = "^0.30.6"
langchain-community = "^0.2.16"
langchain-groq = "^0.1.9"
fastembed = "^0.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
